make code happy
